Upgrade/Install: Make some adjustments to the `move_dir()` function:
* Check for direct PHP flle access and only use `rename()` if true.
* Check whether the destination directory was successfully created.
* Clear the working directory so there is internal parity within the function between the results of a successful `rename()` and a fallback to `copy_dir()`.
* Use `move_dir()` in `WP_Upgrader::move_to_temp_backup_dir()` and `::restore_temp_backup()`.

Follow-up to [51815], [51898], [51899], [51902], [52192], [52284].

Props afragen, peterwilsoncc, dd32, SergeyBiryukov.
See #54166, #51857.
Built from https://develop.svn.wordpress.org/trunk@52289


git-svn-id: http://core.svn.wordpress.org/trunk@51881 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-admin/includes/file.php
@@ -1953,20 +1953,39 @@ function copy_dir( $from, $to, $skip_list = array() ) {
  *
  * @global WP_Filesystem_Base $wp_filesystem WordPress filesystem subclass.
  *
- * @param string $from Source directory.
- * @param string $to   Destination directory.
+ * @param string $from        Source directory.
+ * @param string $to          Destination directory.
+ * @param string $working_dir Optional. Remote file source directory.
+ *                            Default empty string.
  * @return true|WP_Error True on success, WP_Error on failure.
  */
-function move_dir( $from, $to ) {
+function move_dir( $from, $to, $working_dir = '' ) {
 	global $wp_filesystem;
 
-	$wp_filesystem->rmdir( $to );
-	if ( @rename( $from, $to ) ) {
-		return true;
+	if ( 'direct' === $wp_filesystem->method ) {
+		$wp_filesystem->rmdir( $to );
+		if ( @rename( $from, $to ) ) {
+			return true;
+		}
 	}
 
-	$wp_filesystem->mkdir( $to );
+	if ( ! $wp_filesystem->is_dir( $to ) ) {
+		if ( ! $wp_filesystem->mkdir( $to, FS_CHMOD_DIR ) ) {
+
+			// Clear the working directory?
+			if ( ! empty( $working_dir ) ) {
+				$wp_filesystem->delete( $working_dir, true );
+			}
+
+			return new WP_Error( 'mkdir_failed_move_dir', __( 'Could not create directory.' ), $to );
+		}
+	}
 	$result = copy_dir( $from, $to );
+
+	// Clear the working directory?
+	if ( ! empty( $working_dir ) ) {
+		$wp_filesystem->delete( $working_dir, true );
+	}
 
 	return $result;
 }
